requirements.txt
langchain
langchain-openai
rich
python-dotenv
langchain_ollama
